+++ lib/src/widgets.dart
@@ -9,6 +9,7 @@ export 'src/widgets/editable_text.dart';
 export 'src/widgets/framework.dart';
 export 'src/widgets/icon.dart';
 export 'src/widgets/icon_data.dart';
+export 'src/widgets/image.dart';
 export 'src/widgets/spacer.dart';
 export 'src/widgets/text.dart';
 export 'src/widgets/navigator.dart';